requirements.txt
# Notion → GitHub Sync Dependencies
# Python 3.11+ required

# Notion API client
notion-client>=2.2.0

# Environment variable management
python-dotenv>=1.0.0

# CLI interface
click>=8.1.0

# HTTP requests (for image downloads)
requests>=2.31.0

# Rich console output
rich>=13.7.0

# Rate limiting
ratelimit>=2.2.1
